+++ src/Authors.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+class Authors extends React.Component {
+    render() {
+      return (
+        <div>
+            <h1>Писатели Беларуси</h1>
+        </div>
+      )
+    }
+}
+
+export default Authors;
